Add new "single" attribut for FTDI devices.
FTDI devices could be selected with a module index (read value of
some pins) or a serial. In some setups there is only one module
present which can be identified with the VID and PID only. The
new attribute "single" checks for exactly one device with the
VID and PID. It throws an error if 0 or more than 1 devices with
this VID and PID are found.

Example for the syntax:
                <FTDI vendor="0x0403" product="0x6010">
                       <Device single="true"/>
                </FTDI>