Control the state of unused pins with "defaultout" and "defaultoe".
Unused pins were forced to input state until now. This is still
the default behaviour, but it can be changed with the "defaultout"
and "defaultoe" attributes now:

 <FTDI vendor="0x0403" product="0x6010">
  <Device single="true" pinMask="fffffffe" defaultout="00000001" defaultoe="00000001"/>
 </FTDI>

Both are 8 digit hex numbers like "pinMask" defining the output and
output enable values for the pins in the form

  AABBCCDD

With

  AA = the value for the AD pins
  BB = the value for the AC pins
  CC = the value for the BD pins
  DD = the value for the BC pins

Please note that only pins which are 0 in the pin mask can be
controlled with defaultout and defaultoe.

Example: Set out=1 and oe=1 for BC0
  pinMask="fffffffe"
  defaultout="00000001"
  defaultoe="00000001"